Refactor legacy menu controls into dedicated files
