+++ src/client/ui/menu.cpp
@@ -19,6 +19,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "ui.hpp"
 #include "MenuItem.h"
 #include "common/files.hpp"
+#include "menu_controls.hpp"
 
 #include <limits.h>
 #include <algorithm>
@@ -157,92 +158,6 @@ static void Menu_UpdateConditionalState(menuFrameWork_t *menu)
 
 static void Menu_DrawGroups(menuFrameWork_t *menu);
 static void Menu_UpdateGroupBounds(menuFrameWork_t *menu);
-
-/*
-===================================================================
-
-ACTION CONTROL
-
-===================================================================
-*/
-
-static void Action_Free(menuAction_t *a)
-{
-    Z_Free(a->generic.name);
-    Z_Free(a->generic.status);
-    Z_Free(a->cmd);
-    Z_Free(a);
-}
-
-/*
-=================
-Action_Init
-=================
-*/
-static void Action_Init(menuAction_t *a)
-{
-    Q_assert(a->generic.name);
-
-    if ((a->generic.uiFlags & UI_CENTER) != UI_CENTER) {
-        a->generic.x += RCOLUMN_OFFSET;
-    }
-
-    a->generic.rect.x = a->generic.x;
-    a->generic.rect.y = a->generic.y;
-    UI_StringDimensions(&a->generic.rect, a->generic.uiFlags, a->generic.name);
-}
-
-
-/*
-===================================================================
-
-STATIC CONTROL
-
-===================================================================
-*/
-
-/*
-=================
-Static_Init
-=================
-*/
-static void Static_Init(menuStatic_t *s)
-{
-    Q_assert(s->generic.name);
-
-    if (!s->maxChars) {
-        s->maxChars = MAX_STRING_CHARS;
-    }
-
-    s->generic.rect.x = s->generic.x;
-    s->generic.rect.y = s->generic.y;
-
-    UI_StringDimensions(&s->generic.rect,
-                        s->generic.uiFlags, s->generic.name);
-}
-
-/*
-===================================================================
-
-BITMAP CONTROL
-
-===================================================================
-*/
-
-static void Bitmap_Free(menuBitmap_t *b)
-{
-    Z_Free(b->generic.status);
-    Z_Free(b->cmd);
-    Z_Free(b);
-}
-
-static void Bitmap_Init(menuBitmap_t *b)
-{
-    b->generic.rect.x = b->generic.x;
-    b->generic.rect.y = b->generic.y;
-    b->generic.rect.width = b->generic.width;
-    b->generic.rect.height = b->generic.height;
-}
 
 /*
 ===================================================================
@@ -394,106 +309,6 @@ static menuSound_t Keybind_Key(menuKeybind_t *k, int key)
     return QMS_NOTHANDLED;
 }
 
-
-/*
-===================================================================
-
-FIELD CONTROL
-
-===================================================================
-*/
-
-static void Field_Push(menuField_t *f)
-{
-    IF_Init(&f->field, f->width, f->width);
-    IF_Replace(&f->field, f->cvar->string);
-}
-
-static void Field_Pop(menuField_t *f)
-{
-    Cvar_SetByVar(f->cvar, f->field.text, FROM_MENU);
-}
-
-static void Field_Free(menuField_t *f)
-{
-    Z_Free(f->generic.name);
-    Z_Free(f->generic.status);
-    Z_Free(f);
-}
-
-/*
-=================
-Field_Init
-=================
-*/
-static void Field_Init(menuField_t *f)
-{
-    int w = f->width * CONCHAR_WIDTH;
-
-    f->generic.uiFlags &= ~(UI_LEFT | UI_RIGHT);
-
-    if (f->generic.name) {
-        f->generic.rect.x = f->generic.x + LCOLUMN_OFFSET;
-        f->generic.rect.y = f->generic.y;
-        UI_StringDimensions(&f->generic.rect,
-                            f->generic.uiFlags | UI_RIGHT, f->generic.name);
-        f->generic.rect.width += (RCOLUMN_OFFSET - LCOLUMN_OFFSET) + w;
-    } else {
-        f->generic.rect.x = f->generic.x - w / 2;
-        f->generic.rect.y = f->generic.y;
-        f->generic.rect.width = w;
-        f->generic.rect.height = CONCHAR_HEIGHT;
-    }
-}
-
-
-static bool Field_TestKey(menuField_t *f, int key)
-{
-    if (f->generic.flags & QMF_NUMBERSONLY) {
-        return Q_isdigit(key) || key == '+' || key == '-' || key == '.';
-    }
-
-    return Q_isprint(key);
-}
-
-/*
-=================
-Field_Key
-=================
-*/
-static int Field_Key(menuField_t *f, int key)
-{
-    if (IF_KeyEvent(&f->field, key)) {
-        return QMS_SILENT;
-    }
-
-    if (Field_TestKey(f, key)) {
-        return QMS_SILENT;
-    }
-
-    return QMS_NOTHANDLED;
-}
-
-/*
-=================
-Field_Char
-=================
-*/
-static int Field_Char(menuField_t *f, int key)
-{
-    bool ret;
-
-    if (!Field_TestKey(f, key)) {
-        return QMS_BEEP;
-    }
-
-    ret = IF_CharEvent(&f->field, key);
-    if (f->generic.change) {
-        f->generic.change(&f->generic);
-    }
-
-    return ret ? QMS_SILENT : QMS_NOTHANDLED;
-}
 
 /*
 ===================================================================
@@ -2353,145 +2168,6 @@ void MenuList_Sort(menuList_t *l, int offset, int (*cmpfunc)(const void *, const
 /*
 ===================================================================
 
-SLIDER CONTROL
-
-===================================================================
-*/
-
-static menuSound_t Slider_DoSlide(menuSlider_t *s, int dir);
-
-static void Slider_Push(menuSlider_t *s)
-{
-    s->modified = false;
-    s->curvalue = Q_circ_clipf(s->cvar->value, s->minvalue, s->maxvalue);
-}
-
-static void Slider_Pop(menuSlider_t *s)
-{
-    if (s->modified) {
-        float val = Q_circ_clipf(s->curvalue, s->minvalue, s->maxvalue);
-        Cvar_SetValue(s->cvar, val, FROM_MENU);
-    }
-}
-
-static void Slider_Free(menuSlider_t *s)
-{
-    Z_Free(s->generic.name);
-    Z_Free(s->generic.status);
-    Z_Free(s);
-}
-
-static void Slider_Init(menuSlider_t *s)
-{
-    int len = strlen(s->generic.name) * CONCHAR_WIDTH;
-
-    s->generic.rect.x = s->generic.x + LCOLUMN_OFFSET - len;
-    s->generic.rect.y = s->generic.y;
-
-    s->generic.rect.width = (RCOLUMN_OFFSET - LCOLUMN_OFFSET) +
-                            len + (SLIDER_RANGE + 2) * CONCHAR_WIDTH;
-    s->generic.rect.height = CONCHAR_HEIGHT;
-}
-
-static menuSound_t Slider_Click(menuSlider_t *s)
-{
-    vrect_t rect;
-    float   pos;
-    int     x;
-
-    pos = Q_clipf((s->curvalue - s->minvalue) / (s->maxvalue - s->minvalue), 0, 1);
-
-    x = CONCHAR_WIDTH + (SLIDER_RANGE - 1) * CONCHAR_WIDTH * pos;
-
-    // click left of thumb
-    rect.x = s->generic.x + RCOLUMN_OFFSET;
-    rect.y = s->generic.y;
-    rect.width = x;
-    rect.height = CONCHAR_HEIGHT;
-    if (UI_CursorInRect(&rect))
-        return Slider_DoSlide(s, -1);
-
-    // click on thumb
-    rect.x = s->generic.x + RCOLUMN_OFFSET + x;
-    rect.y = s->generic.y;
-    rect.width = CONCHAR_WIDTH;
-    rect.height = CONCHAR_HEIGHT;
-    if (UI_CursorInRect(&rect)) {
-        uis.mouseTracker = &s->generic;
-        return QMS_SILENT;
-    }
-
-    // click right of thumb
-    rect.x = s->generic.x + RCOLUMN_OFFSET + x + CONCHAR_WIDTH;
-    rect.y = s->generic.y;
-    rect.width = (SLIDER_RANGE + 1) * CONCHAR_WIDTH - x;
-    rect.height = CONCHAR_HEIGHT;
-    if (UI_CursorInRect(&rect))
-        return Slider_DoSlide(s, 1);
-
-    return QMS_SILENT;
-}
-
-static menuSound_t Slider_MouseMove(menuSlider_t *s)
-{
-    float   pos, value;
-    int     steps;
-
-    if (uis.mouseTracker != &s->generic)
-        return QMS_NOTHANDLED;
-
-    pos = (uis.mouseCoords[0] - (s->generic.x + RCOLUMN_OFFSET + CONCHAR_WIDTH)) * (1.0f / (SLIDER_RANGE * CONCHAR_WIDTH));
-
-    value = Q_clipf(pos, 0, 1) * (s->maxvalue - s->minvalue);
-    steps = Q_rint(value / s->step);
-
-    s->modified = true;
-    s->curvalue = s->minvalue + steps * s->step;
-    return QMS_SILENT;
-}
-
-static menuSound_t Slider_Key(menuSlider_t *s, int key)
-{
-    switch (key) {
-    case K_END:
-        s->modified = true;
-        s->curvalue = s->maxvalue;
-        return QMS_MOVE;
-    case K_HOME:
-        s->modified = true;
-        s->curvalue = s->minvalue;
-        return QMS_MOVE;
-    case K_MOUSE1:
-        return Slider_Click(s);
-    }
-
-    return QMS_NOTHANDLED;
-}
-
-
-/*
-=================
-Slider_DoSlide
-=================
-*/
-static menuSound_t Slider_DoSlide(menuSlider_t *s, int dir)
-{
-    s->modified = true;
-    s->curvalue = Q_circ_clipf(s->curvalue + dir * s->step, s->minvalue, s->maxvalue);
-
-    if (s->generic.change) {
-        menuSound_t sound = s->generic.change(&s->generic);
-        if (sound != QMS_NOTHANDLED) {
-            return sound;
-        }
-    }
-
-    return QMS_SILENT;
-}
-
-/*
-===================================================================
-
 SEPARATOR CONTROL
 
 ===================================================================
@@ -2517,33 +2193,6 @@ static void Separator_Draw(menuSeparator_t *s)
 {
     if (s->generic.name)
         UI_DrawString(s->generic.x, s->generic.y, UI_RIGHT, COLOR_WHITE, s->generic.name);
-}
-
-/*
-===================================================================
-
-SAVEGAME CONTROL
-
-===================================================================
-*/
-
-static void Savegame_Push(menuAction_t *a)
-{
-    char *info;
-
-    Z_Free(a->generic.name);
-
-    info = SV_GetSaveInfo(a->cmd);
-    if (info) {
-        a->generic.name = info;
-        a->generic.flags &= ~QMF_GRAYED;
-    } else {
-        a->generic.name = UI_CopyString("<EMPTY>");
-        if (a->generic.type == MTYPE_LOADGAME)
-            a->generic.flags |= QMF_GRAYED;
-    }
-
-    UI_StringDimensions(&a->generic.rect, a->generic.uiFlags, a->generic.name);
 }
 
 /*
